+++ Tests/EntityFrameworkCore.SqlServer.EF7/EntityFrameworkCore.SqlServer.EF7.Domain/Repositories/PkDataSources/IUserSuppliedDataSourceEntityRepository.cs
@@ -1,0 +1,21 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using EntityFrameworkCore.SqlServer.EF7.Domain.Entities.PkDataSources;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Entities.Repositories.Api.EntityRepositoryInterface", Version = "1.0")]
+
+namespace EntityFrameworkCore.SqlServer.EF7.Domain.Repositories.PkDataSources
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public interface IUserSuppliedDataSourceEntityRepository : IEFRepository<UserSuppliedDataSourceEntity, UserSuppliedDataSourceEntity>
+    {
+        [IntentManaged(Mode.Fully)]
+        Task<UserSuppliedDataSourceEntity?> FindByIdAsync(long id, CancellationToken cancellationToken = default);
+        [IntentManaged(Mode.Fully)]
+        Task<List<UserSuppliedDataSourceEntity>> FindByIdsAsync(long[] ids, CancellationToken cancellationToken = default);
+    }
+}